unique key feature added

+++ StudentDatabase.cpp
@@ -1,7 +1,7 @@
 #include<bits/stdc++.h>
 
 int c = 0;
-char ch;
+bool exists = false;
 
 using namespace std;
 
@@ -47,8 +47,9 @@ void checkRecord(string inp_string, string checkData, int checkByWhat) {
             studName[0] = toupper(studName[0]);
             studSurname[0] = toupper(studSurname[0]);
             
-            cout<<"  Found -> "<<studId<<" "<<studName<<" "<<studSurname<<" "<<studAge<<endl;
             c++;
+            cout<<"  Found "<<c<<" -> "<<studId<<" "<<studName<<" "<<studSurname<<" "<<studAge<<endl;
+            
         }
     }
     else if (checkByWhat == 2) {
@@ -57,20 +58,23 @@ void checkRecord(string inp_string, string checkData, int checkByWhat) {
             studName[0] = toupper(studName[0]);
             studSurname[0] = toupper(studSurname[0]);
             
-            cout<<"  Found -> "<<studId<<" "<<studName<<" "<<studSurname<<" "<<studAge<<endl;
             c++;
+            cout<<"  Found "<<c<<" -> "<<studId<<" "<<studName<<" "<<studSurname<<" "<<studAge<<endl;
+            
         }
     }
     else if (checkByWhat == 3) {
         if(vec.at(0) == checkData) { 
-            cout<<"  Found -> "<<studId<<" "<<studName<<" "<<studSurname<<" "<<studAge<<endl;
             c++;
+            cout<<"  Found "<<c<<" -> "<<studId<<" "<<studName<<" "<<studSurname<<" "<<studAge<<endl;
+            
         }
     }
     else if (checkByWhat == 4) {
         if(vec.at(3) == checkData) {
-            cout<<"  Found -> "<<studId<<" "<<studName<<" "<<studSurname<<" "<<studAge<<endl;
             c++;
+            cout<<"  Found "<<c<<" -> "<<studId<<" "<<studName<<" "<<studSurname<<" "<<studAge<<endl;
+            
         }
     }
 }
@@ -81,9 +85,9 @@ void searchStudentRecord() {
 
     ifstream ReadFile("data.txt", ios :: in);
 
-    cout<<"\n  1. Search by Name"<<endl;
-    cout<<"  2. Search by Surname"<<endl;
-    cout<<"  3. Search by Id"<<endl;
+    cout<<"\n  1. Search by Id"<<endl;
+    cout<<"  2. Search by Name"<<endl;
+    cout<<"  3. Search by Surname"<<endl;
     cout<<"  4. Search by Age"<<endl;
 
     int choice;
@@ -94,6 +98,19 @@ void searchStudentRecord() {
     switch(choice) {
 
         case 1:
+
+            cout<<"\n  Enter the Id of the student -> ";
+            cin>>checkData;
+            cout<<endl;
+
+            while(getline(ReadFile, inp_string)) { 
+                    checkRecord(inp_string, checkData, 3);
+            }
+            
+            break;
+
+        case 2:
+            
             cout<<"\n  Enter the Name of the student -> ";
             cin>>checkData;
             cout<<endl;
@@ -110,10 +127,10 @@ void searchStudentRecord() {
                     
                     checkRecord(inp_string, checkData, 1);
             }
-            
+
             break;
 
-        case 2:
+        case 3:
             
             cout<<"\n  Enter the Surname of the student -> ";
             cin>>checkData;
@@ -130,19 +147,7 @@ void searchStudentRecord() {
                     }
                     
                     checkRecord(inp_string, checkData, 2);
-            }
-
-            break;
-
-        case 3:
-            cout<<"\n  Enter the Id of the student -> ";
-            cin>>checkData;
-            cout<<endl;
-
-            while(getline(ReadFile, inp_string)) { 
-                    checkRecord(inp_string, checkData, 3);
-            }
-                    
+            }      
 
             break;
 
@@ -170,6 +175,36 @@ void searchStudentRecord() {
 
 }
 
+bool IdExists(int Id) {
+    
+    ifstream ReadFile("data.txt", ios :: in);
+
+    string inp_string, checkDelimiter = ";", tempString = "";
+
+    while(getline(ReadFile, inp_string)) {
+
+        tempString = "";
+
+        for(char c : inp_string) {
+            
+            if (c == ';') break;
+            else tempString += c;
+            
+        }
+        
+        if(tempString.length() != 0) {
+            if(stoi(tempString) == Id) {
+                ReadFile.close();
+                return true;
+            }
+        }
+    }
+    
+    ReadFile.close();
+    return false;
+
+}
+
 void addStudentRecord() {
 
     ofstream WriteFile("data.txt", ios :: app);
@@ -177,23 +212,39 @@ void addStudentRecord() {
     string studName, studSurname;
     int studAge, studId;
 
-    cout<<"\nEnter Student's name -> ";
-    cin>>studName;
+    jumpHere:
 
-    cout<<"Enter "<<studName<<"'s surname -> ";
-    cin>>studSurname;
-    
-    cout<<"Enter "<<studName<<"'s age -> ";
-    cin>>studAge;
-
-    cout<<"Enter "<<studName<<"'s roll no -> ";
+    cout<<"\nEnter Student's Id -> ";
     cin>>studId;
 
-    WriteFile << studId << ";" << studName << ";" << studSurname << ";" << studAge << ";" <<endl;
+    if(!IdExists(studId)) {
 
-    cout<<"\nRecord successfully stored into database"<<endl;
+        cout<<"Enter Student's name -> ";
+        cin>>studName;
+        
+        cout<<"Enter "<<studName<<"'s surname -> ";
+        cin>>studSurname;
 
-    WriteFile.close();
+        cout<<"Enter "<<studName<<"'s age -> ";
+        cin>>studAge;
+
+        WriteFile << studId << ";" << studName << ";" << studSurname << ";" << studAge << ";" <<endl;
+
+        cout<<"\nRecord stored into database successfully!!"<<endl;
+
+        exists = false;
+
+        WriteFile.close();
+
+    }
+    else {
+
+        cout<<"Id -> "<<studId<<" already exixts in the database. Please enter another one"<<endl;
+        goto jumpHere;
+
+        WriteFile.close();
+
+    }
 
 }
 
@@ -205,12 +256,74 @@ void deleteStudentRecord() {
 
 void eraseDatabase() {
 
-    ofstream WriteFile("data.txt", ios :: trunc);
+    string choice;
+
+    cout<<"\nAre you sure you want to Erase the entire database??"<<endl;
+    cout<<"\nEnter yes or no -> ";
+    cin>>choice;
+
+    if(choice == "y" || choice == "yes" || choice == "Yes" || choice == "YES") {
+
+        ofstream WriteFile("data.txt", ios :: trunc);
+
+        cout<<"\nDatabase erased successfully!!"<<endl;
+
+        WriteFile.close();
+    }
+
+}
+
+void printRecord(string inp_string) {
+
+    string studName, studSurname, temp = "";
+    int studAge = 0, studId = 0, counter = 0, i = 0;
+    string check_delimiter;
+
+    vector<string> vec;
+
+    while(i <= inp_string.length()-1) {
+
+        check_delimiter = inp_string[i];
+
+        if(check_delimiter == ";") {
+            vec.push_back(temp);
+            temp = "";
+        }
+        else {
+            temp += check_delimiter;
+        }
+
+        i++;
+    }
+
+    studName = vec.at(1);
+    studSurname = vec.at(2);
+    studId = stoi(vec.at(0));
+    studAge = stoi(vec.at(3));
+
+    cout<<studId<<" "<<studName<<" "<<studSurname<<" "<<studAge<<endl<<endl;
     
-    cout<<"\nDatabase erased successfully"<<endl;
+}
 
-    WriteFile.close();
+void printEntireDatabase() {
 
+    ifstream ReadFile("data.txt", ios :: in);
+
+    string printString;
+
+    system("cls");
+
+    cout<<"Printing the entire database....\n\n"<<endl;
+
+
+    while(getline(ReadFile, printString)) {
+    
+        printRecord(printString);
+        
+    }
+
+    ReadFile.close();
+    
 }
 
 
@@ -228,10 +341,11 @@ int main() {
         ifstream MyReadFile;
         MyReadFile.open("data.txt", ios :: in);
 
-        cout<<"1. Search Student's information"<<endl;
+        cout<<"1. Search student's information"<<endl;
         cout<<"2. Add new student record"<<endl;
         cout<<"3. Delete a student record"<<endl;
         cout<<"4. Erase the entire database"<<endl;
+        cout<<"5. Print whole database"<<endl;
         cout<<"0. Exit"<<endl;
         
         int choice;
@@ -245,6 +359,7 @@ int main() {
             break;
 
         case 1:
+            c = 0;
             searchStudentRecord();
             break;
 
@@ -258,6 +373,10 @@ int main() {
 
         case 4:
             eraseDatabase();
+            break;
+
+        case 5:
+            printEntireDatabase();
             break;
         
         default:
